feat(mount): validate and clamp RAM cache to 90% of available memory to prevent OOM; fix(zipfs): add critical lock during extraction to prevent concurrent stream corruption; feat(update): prompt user with MessageBox before opening browser; refactor(admin): move IsAdministrator to separate class

+++ SimpleZipDrive/Models/AppSettings.cs
@@ -12,9 +12,9 @@ public class AppSettings
 
     private static readonly JsonSerializerOptions JsonOptions = new() { WriteIndented = true };
 
-    public int MaxMemoryPerFileMb { get; set; } = 512;
+    public long MaxMemoryPerFileMb { get; set; } = 512;
 
-    public int MaxMemoryPerFileBytes => MaxMemoryPerFileMb * 1024 * 1024;
+    public long MaxMemoryPerFileBytes => MaxMemoryPerFileMb * 1024 * 1024;
 
     public static AppSettings Load()
     {
